Documents/bitspilani/ema/requirements.txt
asgiref==3.2.3
backcall==0.1.0
backports.csv==1.0.7
boto3==1.11.17
botocore==1.14.17
certifi==2019.9.11
chardet==3.0.4
decorator==4.4.1
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.7
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap3==11.1.0
django-csv-export-view==1.0.1
django-datatable==0.3.1
django-easy-pdf==0.1.1
django-extensions==2.2.5
django-formset-js-improved==0.5.0.2
django-import-export==1.2.0
django-jquery-js==3.1.1
django-mysql==3.2.0
django-queryset-csv==1.0.2
django-storages==1.9.1
django-widget-tweaks==1.4.5
docutils==0.15.2
easy-thumbnails==2.6
et-xmlfile==1.0.1
html5lib==1.0.1
idna==2.8
ipython==7.9.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.15.1
jmespath==0.9.4
MarkupPy==1.14
mysqlclient==1.4.5
odfpy==1.4.0
openpyxl==3.0.0
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
Pygments==2.4.2
PyPDF2==1.26.0
python-dateutil==2.8.1
python-magic==0.4.15
pytz==2019.3
PyYAML==5.1.2
reportlab==3.5.32
requests==2.22.0
s3transfer==0.3.3
six==1.13.0
sqlparse==0.3.0
tablib==0.14.0
traitlets==4.3.3
unicodecsv==0.14.1
urllib3==1.25.6
wcwidth==0.1.7
webencodings==0.5.1
xhtml2pdf==0.2.3
xlrd==1.2.0
xlwt==1.3.0
django-bulk-update==2.2.0
celery==4.4.7
django-celery-results==1.2.1
redis
djangorestframework

